webapp/requirements.txt
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
gevent==23.9.1
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
peewee==3.16.3
psycopg2-binary==2.9.9
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==3.0.0
zope.event==5.0
zope.interface==6.1
